+++ lib/src/cache/cache_driver.dart
@@ -1,11 +1,11 @@
 abstract class CacheDriver {
-  Future<void> put(String key, String value, {Duration? duration});
+  Future<void> put(String key, dynamic value, {Duration duration});
 
-  Future<void> forever(String key, String value);
+  Future<void> forever(String key, dynamic value);
 
   Future<void> delete(String key);
 
-  Future<String?> get(String key);
+  Future<dynamic> get(String key, [dynamic defaultValue]);
 
   Future<bool> has(String key);
 }