+++ lib/src/http/request/request_file.dart
@@ -39,6 +39,21 @@ class RequestFile {
     return _getFileSize(await bytes);
   }
 
+  /// Returns the original file name.
+  /// It is extracted from the request from which the file has been uploaded.
+  /// This should not be considered as a safe value to use for a file name on your servers.
+  String get getClientOriginalName => filename;
+
+  /// Returns the original file extension.
+  /// It is extracted from the original file name that was uploaded.
+  /// This should not be considered as a safe value to use for a file name on your servers.
+  String get getClientOriginalExtension => filename.split('.').last;
+
+  /// Returns the file mime type.
+  /// The client mime type is extracted from the request from which the file
+  /// was uploaded, so it should not be considered as a safe value.
+  String get getClientMimeType => filetype;
+
   /// this function will store the file in your project storage folder
   ///
   /// ```
@@ -47,17 +62,17 @@ class RequestFile {
   /// ```
   Future<String> store({required String path, required String filename}) async {
     path = path.endsWith("/") ? path : "$path/";
-    return Storage().put(path, filename, await bytes);
+    return Storage.put(path, filename, await bytes);
   }
 
-  /// this function will upload the file in your project custom folder
+  /// this function will upload the file in your project custom path
   ///
   /// ```
   /// RequestFile image = req.input('image');
   /// String filename = await image.move('/public/images','myImage.jpg');
   /// ```
   Future<String> move({required String path, required String filename}) async {
-    path = sanitizeRoutePath('$path/$filename');
+    path = sanitizeRoutePath('${Directory.current.path}/$path/$filename');
     File file = File(path);
     Directory directory = Directory(file.parent.path);
     if (!directory.existsSync()) {