+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.1.6+1
+
+- Fixed Storage issues
+
+## 0.1.6
+
+- Fixed Websocket bugs
+- Storage Converted Instance Methods to Static Methods
+- Cache Converted Instance Methods to Static Methods
+
+## 0.1.5+1
+
+- Fixed env issues
+
+## 0.1.5
+
+- Added Logger
+- Added env file
+
 ## 0.1.4
 
 - Implemented Throttle middleware